jbd-sprint4-day1-idrissabanli/requirements.txt
amqp==5.0.2
appdirs==1.4.4
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.3.1
defusedxml==0.7.0rc1
distlib==0.3.1
Django==2.2.8
django-celery-beat==2.1.0
django-cors-headers==3.6.0
django-filter==2.4.0
django-flatpages-i18n==1.0.0
django-jet==1.0.8
django-js-asset==1.2.2
django-modeltrans==0.5.0
django-modeltranslation==0.16.1
django-mptt==0.11.0
django-rosetta==0.9.5
django-timezone-field==4.0
djangorestframework==3.12.2
filelock==3.0.12
idna==2.10
importlib-metadata==2.1.1
kombu==5.0.2
Markdown==3.3.3
oauthlib==3.1.0
packaging==20.8
Pillow==8.0.1
pluggy==0.13.1
polib==1.1.0
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
py==1.10.0
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2020.4
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tox==3.20.1
urllib3==1.26.2
vine==5.0.0
virtualenv==20.2.2
wcwidth==0.2.5
zipp==3.4.0
